refactor: moving strategies initialization to main entry file to allow vuln strategies to be used by API + CLI. Add vuln strategy type to nsecure payload (+ typings) which is not used for now.
