+++ src/depWalker.js
@@ -20,7 +20,7 @@ const is = require("@slimio/is");
 
 // Require Internal Dependencies
 const { mergeDependencies, cleanRange, recursiveRmdir, constants } = require("./utils");
-const { getVulnerabilityStrategy } = require("./vulnerabilities/vulnerabilitiesSource");
+const { getVulnerabilityStrategy } = require("./vulnerabilities/vulnerabilitySource");
 const { analyzeDirOrArchiveOnDisk } = require("./tarball");
 const Dependency = require("./dependency.class");
 const applyWarnings = require("./warnings");
@@ -333,8 +333,10 @@ async function depWalker(manifest, options = Object.create(null)) {
     }
 
     // Search for vulnerabilities relatively to the current initialized strategy
-    const { hydrateNodeSecurePayload } = await getVulnerabilityStrategy();
-    await hydrateNodeSecurePayload(payload.dependencies);
+    const vulnStrategy = await getVulnerabilityStrategy();
+    await vulnStrategy.hydrateNodeSecurePayload(payload.dependencies);
+
+    payload.vulnerabilityStrategy = vulnStrategy.type;
 
 
     // We do this because it "seem" impossible to link all dependencies in the first walk.
