+++ src/vulnerabilities/strategies/npm-audit.js
@@ -23,35 +23,40 @@ async function hydrateNodeSecurePayload(dependencies) {
         const { vulnerabilities } = (await arborist.audit()).toJSON();
 
         Object.keys(vulnerabilities).forEach((packageName) => {
+            const packageVulnerabilities = extractPackageVulnsFromSource(vulnerabilities[packageName]);
             const dependenciesVulnerabilities = dependencies.get(packageName).vulnerabilities;
-            const packageVulnerabilities = [];
-
-            const { via: vulnSources } = vulnerabilities[packageName];
-
-            for (const vulnSource of vulnSources) {
-                const {
-                    title, range, id,
-                    module_name: name,
-                    severity, version,
-                    vulnerableVersions
-                } = vulnSource;
-
-                const vulnerability = {
-                    title,
-                    module_name: name,
-                    severity, version,
-                    vulnerableVersions,
-                    range,
-                    id
-                };
-                packageVulnerabilities.push(vulnerability);
-            }
 
             dependenciesVulnerabilities.push(packageVulnerabilities);
         });
     }
     // eslint-disable-next-line no-empty
     catch {}
+}
+
+function extractPackageVulnsFromSource(packageVulnerabilities) {
+    const vulnerabilitiesFromSource = [];
+    const { via: vulnSources } = packageVulnerabilities;
+
+    for (const vulnSource of vulnSources) {
+        const {
+            title, range, id,
+            module_name: name,
+            severity, version,
+            vulnerableVersions
+        } = vulnSource;
+
+        const vulnerability = {
+            title,
+            module_name: name,
+            severity, version,
+            vulnerableVersions,
+            range,
+            id
+        };
+        vulnerabilitiesFromSource.push(vulnerability);
+    }
+
+    return vulnerabilitiesFromSource;
 }
