+++ src/vulnerabilities/vulnerabilitySource.js
