+++ test/vulnerabilities/vulnerabilitySource.js
@@ -1,7 +1,7 @@
 /* eslint-disable max-nested-callbacks */
 "use strict";
 
-const { setVulnerabilityStrategy, getVulnerabilityStrategy } = require("../../src/vulnerabilities/vulnerabilitiesSource");
+const { setVulnerabilityStrategy, getVulnerabilityStrategy } = require("../../src/vulnerabilities/vulnerabilitySource");
 
 const { VULN_MODE_DB_SECURITY_WG, VULN_MODE_NPM_AUDIT } = require("../../src/vulnerabilities/strategies");
 const NPMAuditStrategyModule = require("../../src/vulnerabilities/strategies/npm-audit");
